+++ src/LinkDotNet.NCronJob/Registry/JobDefinition.cs
@@ -20,12 +20,14 @@ internal sealed record JobDefinition(
     /// <summary>
     /// The JobFullName is used as a unique identifier for the job type including anonymous jobs. This helps with concurrency management.
     /// </summary>
-    public string JobFullName => JobName == Type.Name ? Type.FullName ?? JobName : $"{typeof(DynamicJobFactory).Namespace}.{JobName}";
+    public string JobFullName => JobName == Type.Name
+        ? Type.FullName ?? JobName
+        : $"{typeof(DynamicJobFactory).Namespace}.{JobName}";
 
     public int JobExecutionCount => Interlocked.CompareExchange(ref jobExecutionCount, 0, 0);
 
     public void IncrementJobExecutionCount() => Interlocked.Increment(ref jobExecutionCount);
-    
+
     public RetryPolicyAttribute? RetryPolicy => JobPolicyMetadata?.RetryPolicy;
     public SupportsConcurrencyAttribute? ConcurrencyPolicy => JobPolicyMetadata?.ConcurrencyPolicy;
 }